moved weather logic to new util.
included climate docs in chroma query
