+++ chain/llm_chain.py
@@ -1,5 +1,6 @@
 from app.utils.helpers import call_ollama_api
 from app.utils.logger import log
+from app.utils.weather import get_weather_info
 from chain.memory import MemoryManager
 from chain.vector_db import query_chroma_doc
 
@@ -7,9 +8,10 @@ from chain.vector_db import query_chroma_doc
 def generate_response(prompt: str, memory: MemoryManager) -> str:
     try:
         context = query_chroma_doc(prompt)
+        weather_context = get_weather_info(prompt)
         # Step 1: Update system message with vector context if provided
         if context:
-            memory.update_system_message(context)
+            memory.update_system_message(context + weather_context)
 
         # Step 2: Add user message to memory
         memory.add_message(role="user", content=prompt)
