+++ db/chroma/chroma.py
@@ -7,7 +7,7 @@ from app.config import EMBEDDING_MODEL, OLLAMA_BASE_URL
 ollamma_embeddings = OllamaEmbeddings(model=EMBEDDING_MODEL, base_url=OLLAMA_BASE_URL)
 
 # chroma client
-client = chroma.HttpClient(host="chroma", port=8080)
+client = chroma.HttpClient(host="chroma", port=8000)
 
 # chroma DB
 chroma_db = Chroma(
